+++ src/b2bua/sip-bridge/accounts/account-pool.cc
@@ -34,17 +34,12 @@ const auto& kDefaultTemplateString = "{uri}"s;
 
 namespace {
 
-void removeAssociatedAuthInfo(linphone::Core& core, const linphone::Account& account) {
+void removeAuthInfo(linphone::Core& core, const linphone::Account& account) {
 	const auto identityAddress = account.getParams()->getIdentityAddress();
 	const auto accountAuthInfo = core.findAuthInfo("", identityAddress->getUsername(), identityAddress->getDomain());
 	if (!accountAuthInfo) return;
 
 	core.removeAuthInfo(accountAuthInfo);
-}
-
-void removeAccount(linphone::Core& core, const std::shared_ptr<linphone::Account>& account) {
-	removeAssociatedAuthInfo(core, *account);
-	core.removeAccount(account);
 }
 
 } // namespace
@@ -99,7 +94,7 @@ AccountPool::AccountPool(const std::shared_ptr<sofiasip::SuRoot>& suRoot,
 }
 
 void AccountPool::loadAll() {
-	// Abort ongoing update process (if any).
+	// Abort the ongoing update process (if any).
 	mAccountOpsQueue.clear();
 
 	auto& defaultView = mDefaultView.view;
@@ -168,6 +163,8 @@ void AccountPool::applyOperation(const CreateAccount& op) {
 		return;
 	}
 
+	LOGD << "(Create) Account '" << accountDesc.getUri() << "' (alias='" << accountDesc.getAlias() << "')";
+
 	const auto accountParams = mAccountParams->clone();
 	accountParams->setIdentityAddress(address);
 
@@ -181,13 +178,16 @@ void AccountPool::applyOperation(const CreateAccount& op) {
 	if (mCore->addAccount(linphoneAccount) != 0) {
 		const auto uri = linphoneAccount->getParams()->getIdentityAddress();
 		LOGW << "(Create) Adding new account to core failed for URI '" << uri->asString() << "'";
-		removeAccount(*mCore, linphoneAccount);
+		removeAccount(linphoneAccount);
 		return;
 	}
 
 	if (!tryEmplace(newAccount)) {
-		removeAccount(*mCore, linphoneAccount);
+		removeAccount(linphoneAccount);
+		return;
 	}
+
+	LOGD << "(Create) Account successfully created";
 }
 
 void AccountPool::applyOperation(const DeleteAccount& op) {
@@ -199,7 +199,10 @@ void AccountPool::applyOperation(const DeleteAccount& op) {
 
 	const auto& linphoneAccount = accountToDelete->getLinphoneAccount();
 
-	removeAccount(*mCore, linphoneAccount);
+	LOGD << "(Delete) account '" << linphoneAccount->getParams()->getIdentityAddress()->asStringUriOnly()
+	     << "' (alias='" << accountToDelete->getAlias().str() << "')";
+
+	removeAccount(linphoneAccount);
 
 	for (auto& [_, view] : mViews) {
 		auto& [formatter, map] = view;
@@ -214,6 +217,13 @@ void AccountPool::applyOperation(const UpdateAccount& op) {
 		return;
 	}
 
+	auto& linphoneAccountToUpdate = *accountToUpdate->getLinphoneAccount();
+	// Keep the current account identity before overriding it (needed to retrieve the current auth info)
+	const auto oldIdentityAddress = linphoneAccountToUpdate.getParams()->getIdentityAddress();
+
+	LOGD << "(Update) account '" << oldIdentityAddress->asStringUriOnly() << "' (alias='"
+	     << accountToUpdate->getAlias().str() << "')";
+
 	// Find all current bindings to the old account to update them later
 	auto previousBindings = vector<tuple<string, const Formatter&, AccountMap&>>();
 	previousBindings.reserve(mViews.size());
@@ -225,10 +235,6 @@ void AccountPool::applyOperation(const UpdateAccount& op) {
 	// Update account
 	const auto& newParams = op.newDesc;
 	accountToUpdate->setAlias(newParams.getAlias());
-
-	auto& linphoneAccountToUpdate = *accountToUpdate->getLinphoneAccount();
-	// Keep current account identity before overriding it (needed to retrieve the current authentication information)
-	const auto oldIdentityAddress = linphoneAccountToUpdate.getParams()->getIdentityAddress();
 
 	const auto newLinphoneParams = mAccountParams->clone();
 	const auto newIdentityAddress = linphone::Factory::get()->createAddress(newParams.getUri());
@@ -326,6 +332,11 @@ void AccountPool::handleAuthInfo(const config::v2::Account& account,
 
 		mCore->addAuthInfo(authInfo);
 	}
+}
+
+void AccountPool::removeAccount(const std::shared_ptr<linphone::Account>& account) {
+	account->addListener(make_shared<AuthInfoDeleter>(mCore, mLogPrefix));
+	mCore->removeAccount(account);
 }
 
 std::shared_ptr<Account> AccountPool::getAccountRandomly() const {
@@ -537,6 +548,44 @@ void AccountPool::onDisconnect(int status) {
 	if (status != REDIS_OK) {
 		LOGE << "Reconnecting to Redis after being unexpectedly disconnected (status " << status << ") ...";
 	}
+}
+
+void AccountPool::AuthInfoDeleter::onRegistrationStateChanged(const std::shared_ptr<linphone::Account>& account,
+                                                              linphone::RegistrationState state,
+                                                              const std::string& message) {
+	if (account == nullptr) {
+		LOGE << "Received notification of registration state change (" << static_cast<int>(state)
+		     << ") on empty account with message '" << message << "': abort";
+		return;
+	}
+	const auto core = mCore.lock();
+	if (core == nullptr) {
+		LOGD << "Cannot remove auth info, the Linphone::Core is not available anymore (std::weak_ptr is empty)";
+		return;
+	}
+
+	const auto identity = account->getParams()->getIdentityAddress()->asStringUriOnly();
+
+	switch (state) {
+		case linphone::RegistrationState::None:
+		case linphone::RegistrationState::Progress:
+		case linphone::RegistrationState::Ok:
+		case linphone::RegistrationState::Refreshing:
+			return;
+		case linphone::RegistrationState::Cleared: {
+			LOGD << "Account '" << identity << "' unregistered with message '" << message << "': removing auth info";
+		} break;
+		case linphone::RegistrationState::Failed: {
+			LOGD << "Account '" << identity << "' failed to update registration state with message '" << message
+			     << "': removing auth info";
+		} break;
+		default: {
+			LOGW << "Unknown registration state '" << static_cast<int>(state) << "' for account '" << identity
+			     << "' with message '" << message << "': removing auth info";
+		}
+	}
+
+	removeAuthInfo(*core, *account);
 }
 
 } // namespace flexisip::b2bua::bridge